fix: add check for stored token when trying to logout

+++ src/services/user/logout.ts
@@ -3,6 +3,14 @@ import { errorHeplers } from "@utils/errors/errorHelpers";
 
 export const logout = async (userId: number): Promise<ServiceResult> => {
   try {
+    const storedToken = await refreshTokenRepository.getByUserId(userId);
+    if (!storedToken) {
+      return {
+        statusCode: 400,
+        jsonResponse: { message: "Logout failed: user already logged out" },
+      };
+    }
+
     await refreshTokenRepository.deleteByUserId(userId);
 
     return {
